+++ xx_demo/xx_demo_menu/lib/focus.dart
@@ -1,4 +1,4 @@
-part of 'main.dart';
+part of 'menu_demo.dart';
 
 
 class FocusButton extends StatefulWidget {
